+++ src/Pages/SiteLocation/SiteLocations.jsx
@@ -5,6 +5,7 @@ import { Spinner } from 'react-bootstrap';
 import DataTable from 'react-data-table-component';
 import { Link } from 'react-router-dom';
 import AdminSidebarNav from '../../Components/Admins/AdminSidebarNav';
+import Swal from "sweetalert2";
 
 const SiteLocations = () => {
     const [siteLocations, setSiteLocations] = useState([]);
@@ -66,26 +67,43 @@ const SiteLocations = () => {
             cell: row => <div>
                 <Link to={`/site-location/`+ row._id} className='btn btn-success me-1'>View</Link>
                 <Link to={`/edit-site-location/`+ row._id} className='btn btn-info me-1'>Edit</Link>
-                <button className='btn btn-warning me-1'>Documents</button>
+                <Link to={`/site-document/`+ row._id} className='btn btn-warning me-1'>Documents</Link>
                 <button className='btn btn-danger' onClick={()=>deleteSiteLocation(row._id)}>Delete</button>
             </div>,
             grow:2,
             center:'yes'
         },
     ];
-    const deleteSiteLocation =async(siteLocationId)=>{
-        setIsLoading(true)
-        const response= await axios.delete(`${process.env.REACT_APP_API_URL}/site-location/`+siteLocationId, { withCredentials: true })
-        if(response){
-            setSuccessMessage("Site location has been deleted successfully.")
-            getSiteLocations()
-            setTimeout(() => {
-                setSuccessMessage()
-            }, 2000)
-        }
+    const deleteSiteLocation = async(siteLocationId)=>{
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You want to delete this site location?",
+            //icon: "warning",
+            dangerMode: true,
+            showCancelButton: true,
+            confirmButtonText: 'Confirm'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(`${process.env.REACT_APP_API_URL}/site-location/`+siteLocationId, { withCredentials: true })
+                        .then(res => {
+                            getSiteLocations()
+                            Swal.fire({
+                                title: "Done!",
+                                text: "Site location Successfully Deleted",
+                                icon: "success",
+                                timer: 2000,
+                                button: false
+                            })
+                        });
+            } else if (
+              result.dismiss === Swal.DismissReason.cancel
+            ) {
+               
+            }
+          })
     }
     return (
-        <div className='site-admins'>
+        <div className='site-Locations'>
             <div className="container-fluid">
                 <div className="row my-5">
                     <div className="col-md-2">
